gridGen gridInCheck Created
Both functions have been created to help generate
the actual Etch-A-Sketch program.

gridGen is what generates the grid, so long as the size
provided is an appropriate size. This size is verified by gridInCheck,
and so long as the passed variable is 1-100 then true shall be
returned by both gridGen and gridInCheck.